requirements.txt
# Core dependencies
fastapi==0.95.1
uvicorn==0.22.0
psycopg2-binary==2.9.6
requests==2.30.0
pydantic==1.10.7

# Scientific stack for AI Patterns
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.3.2

# Optional dependencies
python-dotenv==1.0.0  # For loading environment variables from .env file